Added an option to delete all chords
Right-clicking on the trash icon reveals an option to delete all chords of the progression. Since we want to be able to undo this action even just after loading the page, the state window was updated to also save the initial state when it loads. Also, there was originally just a single dropdown in the timeline controls, so the old one was renamed to stay unambiguous.