Merge landing page header with wiki's
Both headers now use the shared component. This also required a global progress bar, which worked best with nextjs-toploader. Since the Support Us button was removed from the landing page in favor of the new header, an originally mobile-only section is now visible on all devices. The App button on the header is now also shown based on the device.