Improved wiki searching
Searching is now based on the plain text of the MDX document. The search bar also shows a split slug instead of just the title when nested inside the tree. Checking for name consistency was removed to allow SEO-friendly techniques like Introduction to Music Theory to be under the slug /music-theory/introduction.